refactor(discovery): split profiler.go (1342 → 6 files) by concern
Seventh worst-first file-size refactor. DeviceProfiler's methods now
distributed across six concern-focused files in the same package.

  profiler.go             273  DeviceProfiler struct + New + Start/Stop + worker
  profiler_types.go       139  emitted types + constants + vendor name consts
  profiler_config.go      175  ProfilerConfig + defaults + adapter + TLS helper
  profiler_scan.go        352  active probes (ports, HTTP, SNMP, full-MIB)
  profiler_resolve.go     143  DNS / NetBIOS / mDNS name resolution
  profiler_infer.go       216  device-type/icon classifier + portToService
  profiler_accessors.go   123  read-only profile / SNMP / status accessors

All under the 600-line warn budget.

Drive-by cleanups forced by --new-from-rev on relocated lines:

- Reused existing deviceTypeRouter/Switch/Firewall/Printer constants
  and serviceUnknown for the relocated classifier literals.
- Extracted vendorFortinet and webServerApache package-level constants
  for the goconst-flagged literals in the HTTP device matcher.
- Added path-level goconst exclusions for the profiler_infer.go
  portToService map (an intentional literal-reference table) and for
  http/https scheme strings in profiler_scan.go.

Build, vet, lint (--new-from-rev=HEAD~1: 0 issues), discovery tests pass.
